Move code into reachability.py
