+++ vulture/core.py
@@ -1,5 +1,6 @@
 import ast
 from fnmatch import fnmatch, fnmatchcase
+from functools import partial
 from pathlib import Path
 import pkgutil
 import re
@@ -12,6 +13,7 @@ from vulture import noqa
 from vulture import utils
 from vulture.config import InputError, make_config
 from vulture.utils import ExitCode
+from vulture.reachability import Reachability
 
 
 DEFAULT_CONFIDENCE = 60
@@ -212,8 +214,6 @@ class Vulture(ast.NodeVisitor):
         self.defined_vars = get_list("variable")
         self.unreachable_code = get_list("unreachable_code")
 
-        self.no_fall_through_nodes = set()
-
         self.used_names = utils.LoggingSet("name", self.verbose)
 
         self.ignore_names = ignore_names or []
@@ -224,15 +224,14 @@ class Vulture(ast.NodeVisitor):
         self.exit_code = ExitCode.NoDeadCode
         self.noqa_lines = {}
 
+        reporter = partial(self._define, collection=self.unreachable_code)
+        self.reachability = Reachability(reporter=reporter)
+
     def scan(self, code, filename=""):
         filename = Path(filename)
         self.code = code.splitlines()
         self.noqa_lines = noqa.parse_noqa(self.code)
         self.filename = filename
-
-        # We can reset the fall_through_nodes for every module to reduce
-        # memory usage:
-        self.no_fall_through_nodes = set()
 
         def handle_syntax_error(e):
             text = f' at "{e.text.strip()}"' if e.text else ""
@@ -263,6 +262,10 @@ class Vulture(ast.NodeVisitor):
                 self.visit(node)
             except SyntaxError as err:
                 handle_syntax_error(err)
+
+        # Reset the reachability internals for every module to reduce memory
+        # usage.
+        self.reachability.reset()
 
     def scavenge(self, paths, exclude=None):
         def prepare_pattern(pattern):
@@ -636,11 +639,8 @@ class Vulture(ast.NodeVisitor):
                 self.defined_funcs, node.name, node, ignore=_ignore_function
             )
 
-        self._handle_reachability_functiondef(node)
-
     def visit_If(self, node):
         self._handle_conditional_node(node, "if")
-        self._handle_reachability_if(node)
 
     def visit_IfExp(self, node):
         self._handle_conditional_node(node, "ternary")
@@ -670,144 +670,17 @@ class Vulture(ast.NodeVisitor):
 
     def visit_While(self, node):
         self._handle_conditional_node(node, "while")
-        self._handle_reachability_while(node)
 
     def visit_MatchClass(self, node):
         for kwd_attr in node.kwd_attrs:
             self.used_names.add(kwd_attr)
 
-    def visit_Try(self, node):
-        self._handle_reachability_try(node)
-
-    def visit_AsyncFor(self, node):
-        self.visit_For(node)
-
-    def visit_For(self, node):
-        self._handle_reachability_for(node)
-
-    def visit_AsyncWith(self, node):
-        return self.visit_With(node)
-
-    def visit_With(self, node):
-        self._handle_reachability_with(node)
-
-    def visit_Module(self, node):
-        self._handle_reachability_module(node)
-
-    def visit_Return(self, node):
-        self._mark_as_no_fall_through(node)
-
-    def visit_Raise(self, node):
-        self._mark_as_no_fall_through(node)
-
-    def visit_Continue(self, node):
-        self._mark_as_no_fall_through(node)
-
-    def visit_Break(self, node):
-        self._mark_as_no_fall_through(node)
-
-    def _can_fall_through(self, node):
-        return node not in self.no_fall_through_nodes
-
-    def _mark_as_no_fall_through(self, node):
-        self.no_fall_through_nodes.add(node)
-
-    def _can_fall_through_statements_analysis(self, statements):
-        """Report unreachable statements.
-        Returns True if we cannot fall though the list of statements
-        """
-        for idx, statement in enumerate(statements):
-            if not self._can_fall_through(statement):
-                try:
-                    next_sibling = statements[idx + 1]
-                except IndexError:
-                    next_sibling = None
-                if next_sibling is not None:
-                    class_name = statement.__class__.__name__.lower()
-                    self._define(
-                        self.unreachable_code,
-                        class_name,
-                        next_sibling,
-                        last_node=statements[-1],
-                        message=f"unreachable code after '{class_name}'",
-                        confidence=100,
-                    )
-                return False
-        return True
-
-    def _handle_reachability_if(self, node):
-
-        has_else = bool(node.orelse)
-
-        if not has_else:
-            if_can_fall_through = self._can_fall_through_statements_analysis(
-                node.body
-            )
-            else_can_fall_through = True
-        else:
-            if_can_fall_through = self._can_fall_through_statements_analysis(
-                node.body
-            )
-            else_can_fall_through = self._can_fall_through_statements_analysis(
-                node.orelse
-            )
-
-        statement_can_fall_through = (
-            if_can_fall_through or else_can_fall_through
-        )
-
-        if not statement_can_fall_through:
-            self._mark_as_no_fall_through(node)
-
-    def _handle_reachability_try(self, node):
-
-        has_else = bool(node.orelse)
-
-        try_can_fall_through = self._can_fall_through_statements_analysis(
-            node.body
-        )
-
-        if not try_can_fall_through and has_else:
-            else_body = node.orelse
-            self._define(
-                self.unreachable_code,
-                "else",
-                else_body[0],
-                last_node=else_body[-1],
-                message="unreachable 'else' block",
-                confidence=100,
-            )
-
-        any_except_can_fall_through = any(
-            self._can_fall_through_statements_analysis(handler.body)
-            for handler in node.handlers
-        )
-
-        statement_can_fall_through = (
-            try_can_fall_through or any_except_can_fall_through
-        )
-
-        if not statement_can_fall_through:
-            self._mark_as_no_fall_through(node)
-
-    def _handle_reachability_for(self, node):
-        self._can_fall_through_statements_analysis(node.body)
-
-    def _handle_reachability_while(self, node):
-        self._can_fall_through_statements_analysis(node.body)
-
-    def _handle_reachability_with(self, node):
-        self._can_fall_through_statements_analysis(node.body)
-
-    def _handle_reachability_functiondef(self, node):
-        self._can_fall_through_statements_analysis(node.body)
-
-    def _handle_reachability_module(self, node):
-        self._can_fall_through_statements_analysis(node.body)
-
     def visit(self, node):
 
+        # Visit children nodes first to allow recursive reachability analysis.
         self.generic_visit(node)
+
+        self.reachability.visit(node)
 
         method = "visit_" + node.__class__.__name__
         visitor = getattr(self, method, None)
