+++ vulture/whitelists/ast_whitelist.py
@@ -13,13 +13,11 @@ whitelist_node_visitor.visit_AugAssign
 whitelist_node_visitor.visit_Await
 whitelist_node_visitor.visit_BinOp
 whitelist_node_visitor.visit_BoolOp
-whitelist_node_visitor.visit_Break
 whitelist_node_visitor.visit_Bytes
 whitelist_node_visitor.visit_Call
 whitelist_node_visitor.visit_ClassDef
 whitelist_node_visitor.visit_Compare
 whitelist_node_visitor.visit_Constant
-whitelist_node_visitor.visit_Continue
 whitelist_node_visitor.visit_Delete
 whitelist_node_visitor.visit_Dict
 whitelist_node_visitor.visit_DictComp
